Changes for metadata fields, Params to Args
Made changes to add metadata to MPilot script function arguments. Changed the term parameter to argument throughout code, so now the one and only way to refer to arguments in an MPilot script is argument.

Note: MPilotEEMSOnlineGraphicsLib.py needs to be merged with the one in the main branch. All other files will hopefully work as replacements in the main branch.