Change default datatype on READ input fields from Float to None. Allows CVTTOFUZZYCAT() to work.
Previously, all input fields to the READ command were defaulting to
"Float". When the resulting field was then used in a CVTTOFUZZYCAT(), it
was throwing an error because the CVTTOFUZZYCAT() command requires
integers as inputs. The data type of a field cannot be explicitly
defined when creating a model in EEMS 2.0 in ArcGIS. Consequently a
model couldn't be uploaded to EEMS Online if a CVTTOFUZZYCAT() command
was used.